linux: Add crypto.c and no-crypto.c with shims for methods implemented in linux.c Adds compatibility shims for linux.c methods when building on Windows.
